requirements.txt
jitsuin-archivist==0.9.4
